+++ src/main/java/uk/gov/hmcts/darts/arm/service/UnstructuredTranscriptionAndAnnotationDeleterProcessor.java
@@ -3,7 +3,7 @@ package uk.gov.hmcts.darts.arm.service;
 import java.util.List;
 
 public interface UnstructuredTranscriptionAndAnnotationDeleterProcessor {
-    List<Integer> markForDeletion(int weeksBeforeCurrentDateInUnstructured, int hoursBeforeCurrentDateInArm, int batchSize);
+    List<Long> markForDeletion(int weeksBeforeCurrentDateInUnstructured, int hoursBeforeCurrentDateInArm, int batchSize);
 
-    List<Integer> markForDeletion(int batchSize);
+    List<Long> markForDeletion(int batchSize);
 }
