+++ src/main/java/uk/gov/hmcts/darts/datamanagement/service/InboundTranscriptionAnnotationDeleterProcessor.java
@@ -3,7 +3,7 @@ package uk.gov.hmcts.darts.datamanagement.service;
 import java.util.List;
 
 public interface InboundTranscriptionAnnotationDeleterProcessor {
-    List<Integer> markForDeletion(int hourThreshold, int batchSize);
+    List<Long> markForDeletion(int hourThreshold, int batchSize);
 
-    List<Integer> markForDeletion(int batchSize);
+    List<Long> markForDeletion(int batchSize);
 }
